organize charts into folder
