+++ src/components/charts/acute-covid-chart.tsx
