+++ src/components/charts/long-covid-chart.tsx
